refactor(io): move inline unit tests into tests.rs
The baseline wants unit tests in `foo/tests.rs` beside `foo/mod.rs` rather than an
inline `#[cfg(test)] mod tests { .. }` block, so module files stay readable. Module
paths are unchanged, so nothing that names these modules moves.

6 test file(s) split out. No test was added, removed or edited -- only relocated.
